Introduce broker-common module
- Move data types to the common module.
