+++ modules/broker-amqp/src/main/java/org/wso2/broker/amqp/AmqpConsumer.java
@@ -24,10 +24,10 @@ import io.netty.channel.ChannelFutureListener;
 import io.netty.channel.ChannelHandlerContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.wso2.broker.amqp.codec.data.ShortString;
 import org.wso2.broker.amqp.codec.frames.BasicDeliver;
 import org.wso2.broker.amqp.codec.frames.ContentFrame;
 import org.wso2.broker.amqp.codec.frames.HeaderFrame;
+import org.wso2.broker.common.data.types.ShortString;
 import org.wso2.broker.core.Consumer;
 import org.wso2.broker.core.ContentChunk;
 import org.wso2.broker.core.Message;
