+++ modules/broker-amqp/src/main/java/org/wso2/broker/amqp/codec/AmqpChannel.java
@@ -22,7 +22,7 @@ package org.wso2.broker.amqp.codec;
 import io.netty.channel.ChannelHandlerContext;
 import org.wso2.broker.amqp.AmqpConsumer;
 import org.wso2.broker.amqp.AmqpException;
-import org.wso2.broker.amqp.codec.data.ShortString;
+import org.wso2.broker.common.data.types.ShortString;
 import org.wso2.broker.core.Broker;
 import org.wso2.broker.core.BrokerException;
 import org.wso2.broker.core.Consumer;
