+++ modules/broker-amqp/src/main/java/org/wso2/broker/amqp/codec/frames/BasicConsumeOk.java
@@ -24,7 +24,7 @@ import io.netty.channel.ChannelHandlerContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.broker.amqp.codec.AmqpConnectionHandler;
-import org.wso2.broker.amqp.codec.data.ShortString;
+import org.wso2.broker.common.data.types.ShortString;
 
 /**
  * AMQP frame for basic.consume
