+++ modules/broker-amqp/src/main/java/org/wso2/broker/amqp/codec/frames/BasicPublish.java
@@ -26,7 +26,7 @@ import org.slf4j.LoggerFactory;
 import org.wso2.broker.amqp.codec.AmqpChannel;
 import org.wso2.broker.amqp.codec.AmqpConnectionHandler;
 import org.wso2.broker.amqp.codec.InMemoryMessageAggregator;
-import org.wso2.broker.amqp.codec.data.ShortString;
+import org.wso2.broker.common.data.types.ShortString;
 
 /**
  * AMQP frame for basic.publish
