+++ modules/broker-amqp/src/main/java/org/wso2/broker/amqp/codec/frames/ConnectionStart.java
@@ -22,8 +22,8 @@ package org.wso2.broker.amqp.codec.frames;
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelHandlerContext;
 import org.wso2.broker.amqp.codec.AmqpConnectionHandler;
-import org.wso2.broker.amqp.codec.data.FieldTable;
-import org.wso2.broker.amqp.codec.data.LongString;
+import org.wso2.broker.common.data.types.FieldTable;
+import org.wso2.broker.common.data.types.LongString;
 
 /**
  * AMQP connection.start frame.
