+++ modules/broker-amqp/src/main/java/org/wso2/broker/amqp/codec/frames/HeaderFrame.java
@@ -24,9 +24,9 @@ import io.netty.channel.ChannelHandlerContext;
 import org.wso2.broker.amqp.codec.AmqpChannel;
 import org.wso2.broker.amqp.codec.AmqpConnectionHandler;
 import org.wso2.broker.amqp.codec.InMemoryMessageAggregator;
-import org.wso2.broker.amqp.codec.data.EncodableData;
-import org.wso2.broker.amqp.codec.data.FieldTable;
-import org.wso2.broker.amqp.codec.data.ShortString;
+import org.wso2.broker.common.data.types.EncodableData;
+import org.wso2.broker.common.data.types.FieldTable;
+import org.wso2.broker.common.data.types.ShortString;
 
 /**
  * AMQP header frame
