+++ modules/broker-amqp/src/main/java/org/wso2/broker/amqp/codec/frames/QueueDeclare.java
@@ -27,8 +27,8 @@ import org.wso2.broker.amqp.codec.AmqpChannel;
 import org.wso2.broker.amqp.codec.AmqpConnectionHandler;
 import org.wso2.broker.amqp.codec.BlockingTask;
 import org.wso2.broker.amqp.codec.ChannelException;
-import org.wso2.broker.amqp.codec.data.FieldTable;
-import org.wso2.broker.amqp.codec.data.ShortString;
+import org.wso2.broker.common.data.types.FieldTable;
+import org.wso2.broker.common.data.types.ShortString;
 import org.wso2.broker.core.BrokerException;
 
 /**
