+++ modules/broker-common/src/main/java/org/wso2/broker/common/data/types/FieldValue.java
@@ -11,13 +11,13 @@
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
+ * KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations
  * under the License.
  *
  */
 
-package org.wso2.broker.amqp.codec.data;
+package org.wso2.broker.common.data.types;
 
 import io.netty.buffer.ByteBuf;
