+++ modules/broker-common/src/main/java/org/wso2/broker/common/data/types/LongString.java
@@ -11,13 +11,13 @@
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
+ * KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations
  * under the License.
  *
  */
 
-package org.wso2.broker.amqp.codec.data;
+package org.wso2.broker.common.data.types;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.netty.buffer.ByteBuf;
@@ -26,57 +26,51 @@ import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 
 /**
- * AMQP Short String.
+ * AMQP Long String.
  *
- * short-string = OCTET *string-char ; length + content
+ * long-string = long-uint *OCTET ; length + content
  */
-public class ShortString implements EncodableData {
-    private static final int MAX_LENGTH = 0xFF;
+public class LongString implements EncodableData {
     private final long length;
-
-    // TODO maybe we should keep a char sequence instead of a byte array.
-    // We should do a perf test before switching to char sequence.
     private final byte[] content;
 
-    public ShortString(long length, char[] content) {
-        this.length = length;
-        this.content = new String(content).getBytes(StandardCharsets.UTF_8);
-    }
-
     @SuppressFBWarnings("EI_EXPOSE_REP2")
-    public ShortString(long length, byte[] content) {
+    public LongString(long length, byte[] content) {
         this.length = length;
         this.content = content;
     }
 
-    public long getSize() {
-        return length + 1;
-    }
-
-    public void write(ByteBuf buf) {
-        buf.writeByte((int) length);
-        buf.writeBytes(content);
-    }
-
-    public static ShortString parse(ByteBuf buf) {
-        int size = buf.readUnsignedByte();
-        byte[] data = new byte[size];
-        buf.readBytes(data);
-
-        return new ShortString(size, data);
-    }
-
-    public static ShortString parseString(String data) {
-        return new ShortString(data.length(), data.getBytes(StandardCharsets.UTF_8));
+    public static LongString parseString(String data) {
+        return new LongString(data.length(), data.getBytes(StandardCharsets.UTF_8));
     }
 
     @Override
-    public String toString() {
-        return new String(content, StandardCharsets.UTF_8);
+    public long getSize() {
+        // We need to add bytes used for size as well
+        return 4 + length;
+    }
+
+    @Override
+    public void write(ByteBuf buf) {
+        buf.writeInt((int) length);
+        buf.writeBytes(content);
+    }
+
+    public static LongString parse(ByteBuf buf) throws Exception {
+        int size = (int) buf.readUnsignedInt();
+
+        if (size < 0) {
+            throw new Exception("Invalid string length");
+        }
+
+        byte[] data = new byte[size];
+        buf.readBytes(data);
+
+        return new LongString(size, data);
     }
 
     /**
-     * Compares {@link ShortString} underlying byte array content.
+     * Compares {@link LongString} underlying byte array content.
      *
      * @param obj {@link Object} to compare with
      * @return True if underlying content is equal
@@ -86,7 +80,7 @@ public class ShortString implements EncodableData {
         if (this == obj) {
             return true;
         }
-        return (obj instanceof ShortString) && (Arrays.equals(content, ((ShortString) obj).content));
+        return (obj instanceof LongString) && (Arrays.equals(content, ((LongString) obj).content));
     }
 
     /**
